wip on function and predicate body conversion and primitive operations

+++ example/GCD.java
@@ -9,46 +9,46 @@ import static byteback.annotations.Special.*;
 public class GCD {
 
 	@Pure
-	public static int gcd_recursive(int a, int b) {
-		return conditional(eq(a, b), a, conditional(gt(a, b), gcd_recursive(a - b, b), gcd_recursive(a, b - a)));
+	public static short gcd_recursive(short a, short b) {
+		return conditional(eq(a, b), a, conditional(gt(a, b), gcd_recursive((short) (a - b), b), gcd_recursive(a, (short) (b - a))));
 	}
 
 	@Predicate
-	public static boolean result_is_gcd(int a, int b, int r) {
+	public static boolean result_is_gcd(short a, short b, short r) {
 		return implies(not(arguments_are_negative(a, b)), eq(r, gcd_recursive(a, b)));
 	}
 
 	@Pure
 	@Predicate
-	public static boolean arguments_are_negative(int a, int b) {
+	public static boolean arguments_are_negative(short a, short b) {
 		return lte(a, 0) | lte(b, 0);
 	}
 
 	@Pure
 	@Predicate
-	public static boolean arguments_are_positive(int a, int b) {
+	public static boolean arguments_are_positive(short a, short b) {
 		return lte(a, 0) & lte(b, 0);
 	}
 
 	@Raise(exception = IllegalArgumentException.class, when = "arguments_are_negative")
 	@Ensure("result_is_gcd")
-	public static int gcd(final int a, final int b) {
+	public static short gcd(final short a, final short b) {
 
 		if (a <= 0 || b <= 0) {
 			throw new IllegalArgumentException("Both arguments must be positive");
 		}
 
-		int r = a;
-		int x = b;
+		short r = a;
+		short x = b;
 
 		while (r != x) {
 			invariant(gt(r, 0) & gt(x, 0));
 			invariant(eq(gcd_recursive(r, x), gcd_recursive(a, b)));
 
 			if (r > x) {
-				r = r - x;
+				r = (short) (r - x);
 			} else {
-				x = x - r;
+				x = (short) (x - r);
 			}
 		}
